Update to .net 9.0 & change dependencies
this should be like 4 different commits but i'm bad at coding.

- Update from .net framework 4.7.2 to .net 9.0
Changes made to windows forms means that the UI is now bigger. I at first tried to change the settings back to what they were but it was a pain in the ass, wouldn't be the exact same, and tbh I like the bigger UI.
- Replaced SoulsFormats dependency with SoulsFormatsNEXT and ESDLang
This required a few changes, but most notably is that instead of using the XML documentation Zeditor will now use the newer JSON documentation. I didn't put a ton of effort into making sure everything was working correctly, so please let me know if something that was working before isn't!
- Replace FastColoredTextBox (FCTB) with Vax-FCTB, a fork thats updated for .net 9.0 (has some kinda odd changes that I needed to workaround though)
- Fixed a couple little UI sizing & placement inconsistencies.
- the "Edit State" button now has F2 as a hotkey and allows pressing enter to submit and close the window (because i like that workflow)